fix: Address PR review suggestions
- Complete decompression function to properly handle JSON+gzip format
- Extract shared EVENT_DICT and generateEntryId utilities to avoid duplication
- Improve size estimation using actual compression instead of magic numbers
- Remove unused compressEntry and packBooleans functions
- Fix TypeScript type issues with proper casting
- Eliminate code duplication between compression utils and scanner

Resolves GitHub Copilot AI review feedback from PR #2
